Status : PM in Making
What is takes to jump from a good #productmanager to a great one? 🏹
This is how I structure my day, and my to-dos, the moment I see this, I realize how broad I need to train my mind to stretch, and how I need to juggle between different perspectives. 📈
This is the best part of being a PM! THE jack of all trades 😜 👑

This format sums up what you need to ace to prepare for your PM interview, what kind of questions to expect from each area, and what traits of yours are judged in each category! 🔝
Hope this helps all the aspiring PMs who are currently preparing for interviews 🎯
When you see this first thing in the morning - you know what to prepare and where your life should circle around 🔂